refactor: floatingButton 수정사항 반영
