+++ packages/ui/src/assets/FloatingButton/Email/index.tsx
@@ -1,8 +1,8 @@
 import { css } from '@emotion/react';
 
 export const Email = ({
-  width = 30,
-  height = 30,
+  width = 26,
+  height = 26,
 }: {
   width?: number;
   height?: number;
@@ -22,20 +22,20 @@ export const Email = ({
       <svg
         width="100%"
         height="100%"
-        viewBox="0 0 30 30"
+        viewBox="0 0 26 26"
         fill="none"
         xmlns="http://www.w3.org/2000/svg"
       >
         <path
-          d="M24.3333 6H5.66667C4.19391 6 3 7.15127 3 8.57143V21.4286C3 22.8487 4.19391 24 5.66667 24H24.3333C25.8061 24 27 22.8487 27 21.4286V8.57143C27 7.15127 25.8061 6 24.3333 6Z"
-          stroke="#FFFFFF"
+          d="M21.089 5.19995H4.91121C3.63482 5.19995 2.6001 6.19772 2.6001 7.42852V18.5714C2.6001 19.8022 3.63482 20.8 4.91121 20.8H21.089C22.3654 20.8 23.4001 19.8022 23.4001 18.5714V7.42852C23.4001 6.19772 22.3654 5.19995 21.089 5.19995Z"
+          stroke="white"
           stroke-width="3"
           stroke-linecap="round"
           stroke-linejoin="round"
         />
         <path
-          d="M3 9L15 16L27 9"
-          stroke="#FFFFFF"
+          d="M2.6001 7.80005L13.0001 13.8667L23.4001 7.80005"
+          stroke="white"
           stroke-width="3"
           stroke-linecap="round"
           stroke-linejoin="round"
