+++ packages/ui/src/assets/FloatingButton/Instagram/index.tsx
@@ -1,8 +1,8 @@
 import { css } from '@emotion/react';
 
 export const Instagram = ({
-  width = 30,
-  height = 30,
+  width = 26,
+  height = 26,
 }: {
   width?: number;
   height?: number;
@@ -22,27 +22,27 @@ export const Instagram = ({
       <svg
         width="100%"
         height="100%"
-        viewBox="0 0 30 30"
+        viewBox="0 0 26 26"
         fill="none"
         xmlns="http://www.w3.org/2000/svg"
       >
         <path
-          d="M21 3H9C5.68629 3 3 5.68629 3 9V21C3 24.3137 5.68629 27 9 27H21C24.3137 27 27 24.3137 27 21V9C27 5.68629 24.3137 3 21 3Z"
-          stroke="#FFFFFF"
+          d="M18.2001 2.59998H7.8001C4.92822 2.59998 2.6001 4.92809 2.6001 7.79998V18.2C2.6001 21.0719 4.92822 23.4 7.8001 23.4H18.2001C21.072 23.4 23.4001 21.0719 23.4001 18.2V7.79998C23.4001 4.92809 21.072 2.59998 18.2001 2.59998Z"
+          stroke="white"
           stroke-width="3"
           stroke-linecap="round"
           stroke-linejoin="round"
         />
         <path
-          d="M15 20C17.7614 20 20 17.7614 20 15C20 12.2386 17.7614 10 15 10C12.2386 10 10 12.2386 10 15C10 17.7614 12.2386 20 15 20Z"
-          stroke="#FFFFFF"
+          d="M12.9998 17.3333C15.3931 17.3333 17.3332 15.3932 17.3332 13C17.3332 10.6067 15.3931 8.66663 12.9998 8.66663C10.6066 8.66663 8.6665 10.6067 8.6665 13C8.6665 15.3932 10.6066 17.3333 12.9998 17.3333Z"
+          stroke="white"
           stroke-width="3"
           stroke-linecap="round"
           stroke-linejoin="round"
         />
         <path
-          d="M22 8V8.001"
-          stroke="#FFFFFF"
+          d="M19.0669 6.93335V6.93435"
+          stroke="white"
           stroke-width="3"
           stroke-linecap="round"
           stroke-linejoin="round"
