+++ packages/ui/src/components/FloatingButton/index.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { theme } from '../../styles';
-import { Email } from '../../assets/Email';
-import { Instagram } from '../../assets/Instagram';
+import { Email } from '../../assets/FloatingButton/Email';
+import { Instagram } from '../../assets/FloatingButton/Instagram';
 
 export interface FloatingProps {
   direction?: string;
@@ -26,7 +26,7 @@ export const FloatingButton = (props: FloatingProps) => {
 const FloatingBtnContainer = styled.div<{ direction?: string }>`
   display: flex;
   flex-direction: ${({ direction }) => (direction ? `${direction}` : 'column')};
-  gap: 16px;
+  gap: 12px;
 
   // 이미지 및 텍스트 드래그 방지
   -webkit-user-select: none;
@@ -35,8 +35,8 @@ const FloatingBtnContainer = styled.div<{ direction?: string }>`
   user-select: none;
 `;
 const FloatingBtnCircle = styled.a`
-  width: 60px;
-  height: 60px;
+  width: 50px;
+  height: 50px;
   background-color: ${theme.palette.Yellow};
   border-radius: 50%;
   box-shadow: ${theme.shadow.Button.Yellow};
@@ -44,9 +44,4 @@ const FloatingBtnCircle = styled.a`
   display: flex;
   align-items: center;
   justify-content: center;
-
-  @media (max-width: 1023px) {
-    width: 50px;
-    height: 50px;
-  }
 `;
